Remove unnecessary constants and move ModifierConstants to Drive file - JR
