+++ src/main/java/org/team157/robot/BuildConstants.java
@@ -5,12 +5,12 @@ public final class BuildConstants {
     public static final String MAVEN_GROUP = "";
     public static final String MAVEN_NAME = "FRC2026";
     public static final String VERSION = "unspecified";
-    public static final int GIT_REVISION = 460;
-    public static final String GIT_SHA = "45b76112dba23d794e7cde47a6e4d6e09bb7f405";
-    public static final String GIT_DATE = "2026-06-22 19:12:47 EDT";
-    public static final String GIT_BRANCH = "feature/anti-slime";
-    public static final String BUILD_DATE = "2026-06-22 19:19:06 EDT";
-    public static final long BUILD_UNIX_TIME = 1782170346578L;
+    public static final int GIT_REVISION = 462;
+    public static final String GIT_SHA = "09a2b1612d14a57e740214f99805b0b4ccd064e9";
+    public static final String GIT_DATE = "2026-06-22 19:35:00 EDT";
+    public static final String GIT_BRANCH = "constantsCleanup";
+    public static final String BUILD_DATE = "2026-06-30 19:57:22 EDT";
+    public static final long BUILD_UNIX_TIME = 1782863842165L;
     public static final int DIRTY = 1;
 
     private BuildConstants() {}
