+++ src/main/java/org/team157/robot/RobotContainer.java
@@ -22,14 +22,14 @@ import edu.wpi.first.wpilibj2.command.sysid.SysIdRoutine;
 import org.littletonrobotics.junction.networktables.LoggedDashboardChooser;
 import org.team157.robot.Constants.FieldConstants;
 import org.team157.robot.Constants.Mode;
-import org.team157.robot.Constants.ModifierConstants;
-import org.team157.robot.Constants.ModifierConstants.DriveControlMode;
 import org.team157.robot.commands.DriveCommands;
 import org.team157.robot.generated.TunerConstants;
 import org.team157.robot.subsystems.HubTimer;
 import org.team157.robot.subsystems.LEDs;
 import org.team157.robot.subsystems.SunstoneMechanism3D;
 import org.team157.robot.subsystems.drive.Drive;
+import org.team157.robot.subsystems.drive.Drive.ModifierConstants;
+import org.team157.robot.subsystems.drive.Drive.ModifierConstants.DriveControlMode;
 import org.team157.robot.subsystems.drive.GyroIO;
 import org.team157.robot.subsystems.drive.GyroIOPigeon2;
 import org.team157.robot.subsystems.drive.ModuleIO;
