+++ src/main/java/org/team157/robot/commands/DriveCommands.java
@@ -28,8 +28,8 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.function.DoubleSupplier;
 import java.util.function.Supplier;
-import org.team157.robot.Constants.ModifierConstants;
 import org.team157.robot.subsystems.drive.Drive;
+import org.team157.robot.subsystems.drive.Drive.ModifierConstants;
 
 public class DriveCommands {
     private static final double DEADBAND = 0.2;
